+++ src/commands/manuals/test.rs
@@ -35,6 +35,11 @@ DESCRIPTION
     be specified in a configuration file, which will be used when a target is
     omitted from the command-line.
     
+    When the '--dut' option and the '--tb' option are omitted, all possible 
+    DUTs and TBs are included, leaving the respective DUT and TB environment 
+    variables empty. Filesets using valid string swapping variables such as 
+    'orbit.dut.name' will resolve to the wildcard character ('*').
+    
     If '--list' is used, then it will display a list of the available targets to
     the user. Using '--list' in combination with a target from '--target' will
     display any detailed help information the target has documented in its 
@@ -77,9 +82,6 @@ OPTIONS
 
     --list, -l
         View available targets and exit
-
-    --all
-        Include all hdl files of the current project
 
     --fileset <key=glob>...
         A glob-style pattern identified by name to include in the blueprint