+++ CHANGELOG.md
@@ -19,8 +19,11 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 
 ### Features
 - Adds [Rust security audit](https://github.com/rustsec/rustsec) to audit dependency crates and report vulnerabilities during CI/CD
+- Adds "file" to JSON data of a design unit
 
 ### Changes
+- Default behavior for `orbit build` and `orbit test` when not specifying `--top`, `--dut`, and `--tb` is to include all source files (deprecates `--all` flag) and no longer attempts to auto-detect a top-level, dut, or testbench
+- When using the `--all` flag, filesets using valid string swapping variables such as `orbit.dut.name`, `orbit.tb.name`, and `orbit.top.name`, resolve to the wildcard pattern (`*`)
 - Reverts `repository` field back to `source` to be more consistent with its intended usage and flexible functionality
 
 ### Fixes